requirements.txt
torch>=2.0.0
torchvision>=0.15.0
einops>=0.6.0
lpips>=0.1.4
scipy>=1.10.0
numpy>=1.24.0
Pillow>=9.0.0
matplotlib>=3.7.0
